Add Noticeable icon Widget to the menubar
The call to `noticeable.render(...)` is made in the main app component
that is initialized only once.

We pass options as third parameter.

The `root` option defines from where to search the widget selector.
Since the widget is rendered inside a Web component, the root is the
current shadowRoot.

The `selector` option overrides the selector that is configured with the
live widget editor on the Noticeable dashboard. Passing the selector as
an option is optional and made here to ensure that you get a working
example even if you use this project sample with your own widget without
configuring the selector with the live editor.